Move classes around into respective packages for neatness
Add in logic for receiving orders and new customer at the same time
