+++ src/main/java/ie/atu/week5/customerapp/Customer/CustomerController.java
@@ -1,12 +1,10 @@
-package ie.atu.week5.customerapp;
+package ie.atu.week5.customerapp.Customer;
 
 import jakarta.validation.Valid;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/customers")
