+++ src/main/java/ie/atu/week5/customerapp/Customer/CustomerRepository.java
@@ -1,5 +1,6 @@
-package ie.atu.week5.customerapp;
+package ie.atu.week5.customerapp.Customer;
 
+import ie.atu.week5.customerapp.Customer.Customer;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface CustomerRepository extends MongoRepository<Customer, String> {
