+++ src/main/java/ie/atu/week5/customerapp/CustomerOrder/CustomerOrderRequest.java
@@ -1,5 +1,7 @@
-package ie.atu.week5.customerapp;
+package ie.atu.week5.customerapp.CustomerOrder;
 
+import ie.atu.week5.customerapp.Customer.Customer;
+import ie.atu.week5.customerapp.Order.Order;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
