--- src/main/java/ie/atu/week5/customerapp/CustomerOrderController.java
@@ -1,36 +1,0 @@
-package ie.atu.week5.customerapp;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/api")
-public class CustomerOrderController {
-
-    private final CustomerRepository customerRepository;
-
-    private final OrderRepository orderRepository;
-
-    public CustomerOrderController(CustomerRepository customerRepository, OrderRepository orderRepository) {
-        this.customerRepository = customerRepository;
-        this.orderRepository = orderRepository;
-    }
-
-    @PostMapping("/customer-with-orders")
-    public ResponseEntity<String> createCustomerWithOrders(@RequestBody CustomerOrderRequest customerOrderRequest) {
-
-        // 1. Save the Customer and get the generated customer ID
-
-
-        // 2. Save the Orders and link them to the customer
-
-
-        return ResponseEntity.ok("Customer and orders created successfully");
-    }
-}
